ci: record-if-missing when coming from stainless (#4408)
# What does this PR do?

we will typically need to record the missing json for net new APIs. use
record-if-missing so that the integration tests can re-record and commit
the files to the PR

set the stainless inference mode to record-if-missing, and properly pass
the pr_head_sha on workflow_call.

## Test Plan

see https://github.com/llamastack/llama-stack/actions/runs/20318245671
which uses this commit.

Signed-off-by: Charlie Doern <cdoern@redhat.com>